requirements.txt
idna==2.8
numpy==1.18.1
Pillow==7.0.0
tweepy==3.8.0
urllib3==1.25.8
